+++ src/api/external_app.py
@@ -52,7 +52,7 @@ def create_external_api_app() -> FastAPI:
     app.include_router(api_v1_router, prefix=API_V1_PREFIX, tags=["API v1"])
     app.include_router(domains_router, prefix=API_V1_PREFIX, tags=["Domain"])
     app.include_router(
-        dt_api_router, prefix=API_DIGITALTWIN_PREFIX, tags=["Digital Twin"]
+        dt_api_router, prefix=API_DIGITALTWIN_PREFIX, tags=["Knowledge Graph"]
     )
     app.include_router(graphql_router, prefix=API_GRAPHQL_PREFIX, tags=["GraphQL"])